requirements.txt
# Core dependencies
numpy==1.23.3
diffusers
transformers==4.36.0
opencv-python==4.10.0.84
pycocotools==2.0.8
safetensors==0.4.3
supervision==0.25.0
timm==0.6.12
yapf==0.43.0
ultralytics
segment-anything-fast
flask-cors==4.0.0

# Authentication and database dependencies
flask-sqlalchemy==3.1.1
flask-session==0.5.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0


# Utility dependencies
accelerate
addict==2.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
platformdirs==4.3.6
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.13.1
segment-anything==1.0
six==1.17.0
sympy==1.13.1
tokenizers>=0.14,<0.19  # Adjusted to match transformers' requirements
tomli==2.2.1
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.21.0
